We need consisten resource names so these can't be true custom resources
