+++ .rubocop.yml
@@ -1,3 +1,3 @@
 Lint/AmbiguousOperator:
   Exclude:
-    - 'resources/domain.rb'
+    - 'libraries/domain.rb'
